Add wrapper to read and implement TODO
If a parent is writing to a pipe and a child empties the pipe, the
child read can return before the entire struct is read. The program
would then continue on with a corrupt struct.

Implement a wrapper to guarantee that all bytes are read, barring
any errors. This fix unlocks achievements much more reliably.

write isn't as much of a problem because it will always write
all bytes unless the buffer is full, which isn't generally a problem
here.