Fix signal problems in parent-child communication
Change the parent-child communication to only send one signal
when the parent requests the child to store changes. Doing so is
necessary because having multiple signals in flight is dangerous
and can lead to losing signals (because pending signals are only
indicated with a binary flag). Subsequent changes should still
move actions out of the signal handler.

Instead of multiple signals, only one signal is sent with the
number of changes the child should act on. Then the child
will take appropriate actions based on the number of changes
and also commit the changes and update the parent inherently.

This fixes inconsistent behavior when storing changes to Steam.

Also add some error checking  and remove spurious check that
would give a warning when the achievements list had already
been populated. The achievements list is properly destroyed
before reusing the variable, so there's no need for the check.