Implement async loading of app icons
App icons were not downloaded asynchronously, so on first run
of SamRewritten, the main window would be completely frozen
for a long time (>1 min). So, download app icons asynchronously.

Now the main window is usable (though still a bit stuttery) and the
game rows are presented before before all icons are downloaded.