Fix asynchronous app icon loading
In the downloader, silencing async's future return is a bad idea
because the future is immediately destructed, and the main thread
waits on it, making the program single-threaded.
Instead of async, just use a much simpler C++ detached thread to
go do the work for us. Trying to use gdk_threads_add_idle_full to get
threading functionality didn't really work for offloading computation
from the main thread, but this C++ threading approach works very
well.

Implemenent a semaphore because C++ doesn't have a native
one and use it to limit simultaneous downloads to 10.
Otherwise when tons of downloads launch, network resolution
can start failing.
Also end users with tons of games could potentially launch
thousands of threads, which isn't a great idea.

Use a mutex on updating the main GUI because modifying the GUI
concurrently is dangerous (and pixbuf would start spuriously
failing)

Now in practice, the GUI is very usable during icon loading time.