Implement committing stat changes
Plumb the rest of the logic through the unlock path
Stats are now functional.

Misc improvements:
Invert type checking on CLI because an int could be
interpreted as a float incorrectly. By checking for float
first, we get the right value: an int can be a float but a
float can't be an int.
Fix unitialized valid_conversion causing stat changes to
never show up.
Various coding and style improvements along the way

Fixes #13