Implement launching app in separate window
This is the very last feature that the origincal SAM does that
SamRewritten doesn't (in fact, that's the only way for the
original SAM to work.) Implement it. Now we can do both -
launch in the same window or a separate window.

Get rid of 'cli' variable logic and just return true or false for
CLI mode. Doing so also brings down the complexity of the
go_cli_mode function to only do one action at a time.